update DST generating script to work both for low gain analysis and real data
git-svn-id: https://svnsrv.desy.de/desy/EVNDISP/trunk@2637 766874ec-0625-4018-90f7-740f284a29e6

+++ scripts/VTS/helper_scripts/UTILITY.make_DST_sub.sh
@@ -8,6 +8,7 @@ source $EVNDISPSYS/setObservatory.sh VTS
 RUN=RUNFILE
 PED=PEDESTALS
 SUMW=SUMWINDOW
+LMULT=LLLOWGAIN
 
 # temporary (scratch) directory
 if [[ -n $TMPDIR ]]; then
@@ -19,47 +20,44 @@ echo "Temporary directory: $TEMPDIR"
 mkdir -p $TEMPDIR
 
 # output data files are written to this directory
-ODIR="$VERITAS_USER_DATA_DIR/analysis/EVD400_DST/$SUMW/"
+ODIR=OUTPUTDIR
 mkdir -p $ODIR
 
 # output log files are written to this directory
 LOGDIR="$ODIR"
 mkdir -p $LOGDIR
 
-# check for the existence of a run parameter file corresponding
-# to the given summation window; if it exists, remove it and replace it
-# with the LOWGAIN runparameter file, modified for the given sum window
-RUNPDIR="$VERITAS_EVNDISP_AUX_DIR/ParameterFiles/"
-RUNPFILE="EVNDISP.reconstruction.SW${SUMW}_noDoublePass.runparameter"
-if [ -e "$RUNPDIR/$RUNPFILE" ]; then
-    rm "$RUNPDIR/$RUNPFILE"
-fi
-
-cat $RUNPDIR/EVNDISP.reconstruction.LOWGAIN.runparameter | sed s|"XX"|"$SUMW"|g > $RUNPDIR/$RUNPFILE
-
 # eventdisplay reconstruction parameter
-ACUTS=$RUNPFILE
+ACUTS=RRRRPFILE
 
 #########################################
-# pedestal and tzero calculation
+# pedestal and tzero calculation. No tzeros needed for lmult. 
 if [[ $PED == "1" ]]; then
     rm -f $LOGDIR/$RUN.ped.log
     $EVNDISPSYS/bin/evndisp -runnumber=$RUN -runmode=1 &> $LOGDIR/$RUN.ped.log
-    
-    rm -f $LOGDIR/$RUN.tzero.log
-    $EVNDISPSYS/bin/evndisp -runnumber=$RUN -runmode=7 -nocalibnoproblem &> $LOGDIR/$RUN.tzero.log
+
+    if [[ $LMULT == "0" ]] ; then    
+	rm -f $LOGDIR/$RUN.tzero.log
+    	$EVNDISPSYS/bin/evndisp -runnumber=$RUN -runmode=7 -nocalibnoproblem &> $LOGDIR/$RUN.tzero.log
+    fi
+fi
+
+
+#########################################
+# Other options. hilo runs should have all events analysed.
+if [[ $LMULT == "0" ]] ; then
+	OPT=" -nevents=5000 "
+else
+	OPT=" -lowgaincalibrationfile calibrationlist.LowGain.CALIB.dat "
 fi
 
 #########################################
 # run eventdisplay
 rm -f $LOGDIR/$RUN.log
-$EVNDISPSYS/bin/evndisp -runnumber=$RUN -runmode=4 -nevents=5000 -nocalibnoproblem -reconstructionparameter $ACUTS -dstfile $TEMPDIR/$RUN.DST.root &> $LOGDIR/$RUN.DST.log
+$EVNDISPSYS/bin/evndisp -runnumber=$RUN -runmode=4 -nocalibnoproblem $OPT -reconstructionparameter $ACUTS -dstfile $TEMPDIR/$RUN.DST.root &> $LOGDIR/$RUN.DST.log
 
 # move data file from temp dir to data dir
 cp -f -v $TEMPDIR/$RUN.DST.root $ODIR/$RUN.DST.root
 rm -f $TEMPDIR/$RUN.DST.root
-
-# remove run parameter file
-rm -f $RUNPDIR/$RUNPFILE
 
 exit
